+++ src/Database/Models/ShopItem.cs
@@ -5,11 +5,11 @@ namespace Anantarupa.Database.Models
 {
     public partial class ShopItem
     {
-        public long ShopItemId { get; set; }
-        public long ItemId { get; set; }
-        public long CurrencyType { get; set; }
-        public long Price { get; set; }
-        public long AllowedQuantity { get; set; }
+        public int ShopItemId { get; set; }
+        public int ItemId { get; set; }
+        public int CurrencyType { get; set; }
+        public int Price { get; set; }
+        public int AllowedQuantity { get; set; }
 
         public virtual Currency CurrencyTypeNavigation { get; set; } = null!;
         public virtual Item Item { get; set; } = null!;
